Safari-611.1.10/LayoutTests/crypto/workers/subtle/ecdh-import-key-derive-key-expected.txt

[Worker] Test deriving AES Keys with imported ECDH base key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ecdh-import-key-derive-key.js
PASS [Worker] derivedKey.type is 'secret'
PASS [Worker] derivedKey.extractable is true
PASS [Worker] derivedKey.algorithm.name is 'AES-CBC'
PASS [Worker] derivedKey.algorithm.length is 128
PASS [Worker] derivedKey.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
